+++ src/modules/mo-core/data-transfer-objects/image-upload.dto.ts
@@ -1,0 +1,13 @@
+import { Expose } from 'class-transformer';
+import { IsBoolean, IsString, IsOptional } from 'class-validator';
+
+export class ImageUploadDto {
+  @IsBoolean()
+  @Expose()
+  success: boolean;
+
+  @IsString()
+  @IsOptional()
+  @Expose()
+  imageUrl?: string;
+}