+++ src/modules/mo-project-dto/data-transfer-objects/external-secret-storage/external-secret-store-response.dto.ts
@@ -1,7 +1,7 @@
 import { Expose } from 'class-transformer';
 import { IsOptional, IsString } from 'class-validator';
 
-export class DeleteExternalSecretsStoreResponseDto {
+export class ExternalSecretStoreResponseDto {
   @Expose()
   @IsString()
   status: 'ERROR' | 'SUCCESS';