+++ src/modules/mo-project-dto/data-transfer-objects/external-secret-storage/index.ts
@@ -1,9 +1,6 @@
+export * from './external-secret-store-response.dto';
 export * from './create-external-secret-store-request.dto';
-export * from './create-external-secret-store-response.dto';
-
-export * from './delete-external-secret-store-request.dto';
-export * from './delete-external-secret-store-response.dto';
-
+export * from './list-external-secret-stores-request.dto';
 export * from './list-external-secret-stores-response.dto';
 
 export * from './list-external-secrets-request.dto';